Refactor: 파일 분리 & 핸들러 속성 추가 (#161)
