+++ modules/shared/components/calendar/type.ts
@@ -1,3 +1,5 @@
+import type { DateData } from 'react-native-calendars';
+
 type Selected = {
   color: string;
 };
@@ -5,3 +7,10 @@ type Selected = {
 export type SelectedDates = {
   [key in string]: Selected;
 };
+
+export type DateInfo = DateData | null;
+
+export interface Icalendar {
+  current?: string;
+  onDayPress?: (arg0: DateInfo, arg1: DateInfo) => void;
+}
