+++ modules/shared/utils/markedDate.ts
@@ -2,8 +2,8 @@ import { MarkedDates } from 'react-native-calendars/src/types';
 import _ from 'lodash';
 
 import { getDatesInRange } from '@shared/utils/utils';
-import { SelectedDates } from '../type';
-import { customTextStyle } from '../style';
+import { SelectedDates } from '../components/calendar/type';
+import { customTextStyle } from '../components/calendar/style';
 
 /**
  * 시작 날짜부터 끝 날짜까지 선택된 범위를 캘린더 속성에 맞게 반환한다.
